Move exceptions.ts, create exception-messages.ts for reusable error message definitions
