+++ packages/mds-stream/kafka/stream-producer.ts
@@ -18,10 +18,9 @@ import { Kafka, Producer } from 'kafkajs'
 import { isArray } from 'util'
 import { Nullable } from '@mds-core/mds-types'
 import logger from '@mds-core/mds-logger'
-import { isDefined, ClientDisconnectedError } from '@mds-core/mds-utils'
+import { isDefined, ClientDisconnectedError, INITIALIZE_CLIENT_MESSAGE } from '@mds-core/mds-utils'
 import { StreamProducer } from '../stream-interface'
 import { getKafkaBrokers } from './helpers'
-import { INITIALIZE_CLIENT_MESSAGE } from '../@types'
 
 export interface KafkaStreamProducerOptions {
   clientId: string
